Created a plugin for editing the arg list of nvim
The Plugin open a new buffer and fills the buffer with the contents of
the arglist.
For each argument there is one line in the buffer.

The buffer can then be edited and when closing the buffer with q or ESC
the lines are then stored in the arglist.

This makes it easy to edit the arglist by just removing lines from the
buffer adding lines and editing existing lines.